create a bootable image
Add a post-image script to make use of the host-mkbootimg package and
automatically create a boot.img that can be flashed to the boot partition.

Also make the kernel override u-boot's provided commandline so console
messages can be sent to the screen, and root partition can be set since
the goal is to avoid a ramdisk (the boot partition is only 10MB).